Move styles to settings stylesheet
