--- app/main/plugins/core/cerebro/settings/Settings/components/styles.css
@@ -1,14 +1,0 @@
-.hotkey {
-
-}
-
-.hotkeyInput {
-  font-size: 16px;
-  line-height: 34px;
-  padding: 0 10px;
-  box-sizing: border-box;
-  width: 100%;
-  border-color: #d9d9d9 #ccc #b3b3b3;
-  border-radius: 4px;
-  border: 1px solid #ccc;
-}
